Distinct designs for mesh and clasp w/2pins per side
This design is no good. It works OK in the x/y direction, I can
assemble a cube with no top/bottom. However, to do so you have
to let one of the sides become slightly longer (specificalyl, it
becomes side+2*mesh_height long). This makes the top/bottom
of the cube effectively impossible to attach